feat: extract partners from PayPal
Interesting. They are handling more information on the server so I had
to learn how to use an Offline Document to parse the HTML for the
information. I'm surprised that only three partners were listed.

I can already see that I will need to restructure the way I handle the
extraction to apply different strategies.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>